fix(delivery): close terminal CI integration gaps
